Refactors model into a more generic encapsulated implementation
