+++ client/my-sites/plans-features-main/index.tsx
@@ -13,6 +13,7 @@ import {
 	PLAN_HOSTING_TRIAL_MONTHLY,
 	PLAN_WOOEXPRESS_PLUS,
 	PLAN_ENTERPRISE_GRID_WPCOM,
+	PLAN_FREE,
 } from '@automattic/calypso-products';
 import { Button, Spinner } from '@automattic/components';
 import { WpcomPlansUI } from '@automattic/data-stores';
@@ -46,6 +47,7 @@ import scrollIntoViewport from 'calypso/lib/scroll-into-viewport';
 import { addQueryArgs } from 'calypso/lib/url';
 import PlanNotice from 'calypso/my-sites/plans-features-main/components/plan-notice';
 import PlanTypeSelector from 'calypso/my-sites/plans-features-main/components/plan-type-selector';
+import useIsCustomDomainAllowedOnFreePlan from 'calypso/my-sites/plans-features-main/hooks/use-is-custom-domain-allowed-on-free-plan';
 import { FeaturesGrid, ComparisonGrid } from 'calypso/my-sites/plans-grid';
 import useGridPlans from 'calypso/my-sites/plans-grid/hooks/npm-ready/data-store/use-grid-plans';
 import usePlanFeaturesForGridPlans from 'calypso/my-sites/plans-grid/hooks/npm-ready/data-store/use-plan-features-for-grid-plans';
@@ -60,14 +62,13 @@ import { getCurrentPlan, isCurrentUserCurrentPlanOwner } from 'calypso/state/sit
 import { getSitePlanSlug, getSiteSlug, isCurrentPlanPaid } from 'calypso/state/sites/selectors';
 import useStorageAddOns from '../add-ons/hooks/use-storage-add-ons';
 import ComparisonGridToggle from './components/comparison-grid-toggle';
-import { FreePlanFreeDomainDialog } from './components/free-plan-free-domain-dialog';
-import { FreePlanPaidDomainDialog } from './components/free-plan-paid-domain-dialog';
 import { LoadingPlaceHolder } from './components/loading-placeholder';
+import ModalContainer from './components/plan-upsell-modal';
+import { useProgressBlockingModal } from './components/plan-upsell-modal/hooks/use-progress-blocking-model';
 import usePricedAPIPlans from './hooks/data-store/use-priced-api-plans';
 import usePricingMetaForGridPlans from './hooks/data-store/use-pricing-meta-for-grid-plans';
 import useCurrentPlanManageHref from './hooks/use-current-plan-manage-href';
 import useFilterPlansForPlanFeatures from './hooks/use-filter-plans-for-plan-features';
-import useIsCustomDomainAllowedOnFreePlan from './hooks/use-is-custom-domain-allowed-on-free-plan';
 import useIsPlanUpsellEnabledOnFreeDomain from './hooks/use-is-plan-upsell-enabled-on-free-domain';
 import useObservableForOdie from './hooks/use-observable-for-odie';
 import usePlanBillingPeriod from './hooks/use-plan-billing-period';
@@ -236,8 +237,10 @@ const PlansFeaturesMain = ( {
 	showLegacyStorageFeature = false,
 	isSpotlightOnCurrentPlan,
 }: PlansFeaturesMainProps ) => {
-	const [ isFreePlanPaidDomainDialogOpen, setIsFreePlanPaidDomainDialogOpen ] = useState( false );
-	const [ isFreePlanFreeDomainDialogOpen, setIsFreePlanFreeDomainDialogOpen ] = useState( false );
+	const [ isModalOpen, setIsModalOpen ] = useState( false );
+	const [ lastClickedPlan, setLastClickedPlan ] = useState< string | null >( null );
+	const { resolveDisplayedModal } = useProgressBlockingModal( { paidDomainName } );
+
 	const [ showPlansComparisonGrid, setShowPlansComparisonGrid ] = useState( false );
 	const translate = useTranslate();
 	const storageAddOns = useStorageAddOns( { siteId, isInSignup } );
@@ -308,40 +311,24 @@ const PlansFeaturesMain = ( {
 		return false;
 	};
 
-	const toggleIsFreePlanPaidDomainDialogOpen = useCallback( () => {
-		setIsFreePlanPaidDomainDialogOpen( ! isFreePlanPaidDomainDialogOpen );
-	}, [ isFreePlanPaidDomainDialogOpen ] );
-
 	const handleUpgradeClick = useCallback(
 		( cartItems?: MinimalRequestCartProduct[] | null ) => {
 			const cartItemForPlan = getPlanCartItem( cartItems );
+			const { product_slug: planSlug = PLAN_FREE } = cartItemForPlan ?? {};
+			setLastClickedPlan( planSlug );
+			if ( isFreePlan( planSlug ) ) {
+				recordTracksEvent( 'calypso_signup_free_plan_click' );
+			}
+
+			const shouldDisplayBlockingModal = resolveDisplayedModal( planSlug ) !== null;
+			if ( shouldDisplayBlockingModal ) {
+				setIsModalOpen( true );
+				return;
+			}
+
 			const cartItemForStorageAddOn = cartItems?.find(
 				( items ) => items.product_slug === PRODUCT_1GB_SPACE
 			);
-			const planPath = cartItemForPlan?.product_slug
-				? getPlanPath( cartItemForPlan.product_slug )
-				: '';
-			const checkoutUrlWithArgs = addQueryArgs(
-				{ ...( withDiscount && { coupon: withDiscount } ) },
-				`/checkout/${ siteSlug }/${ planPath }`
-			);
-
-			// `cartItemForPlan` is empty if Free plan is selected. Show `FreePlanPaidDomainDialog`
-			// in that case and exit. `FreePlanPaidDomainDialog` takes over from there.
-			// It only applies to main onboarding flow and the paid media flow at the moment.
-			// Standardizing it or not is TBD; see Automattic/growth-foundations#63 and pdgrnI-2nV-p2#comment-4110 for relevant discussion.
-			if ( ! cartItemForPlan ) {
-				recordTracksEvent( 'calypso_signup_free_plan_click' );
-				if ( paidDomainName ) {
-					toggleIsFreePlanPaidDomainDialogOpen();
-					return;
-				}
-				if ( isPlanUpsellEnabledOnFreeDomain.result ) {
-					setIsFreePlanFreeDomainDialogOpen( true );
-					return;
-				}
-			}
-
 			if ( cartItemForStorageAddOn?.extra ) {
 				recordTracksEvent( 'calypso_signup_storage_add_on_upgrade_click', {
 					add_on_slug: cartItemForStorageAddOn.extra.feature_slug,
@@ -353,16 +340,17 @@ const PlansFeaturesMain = ( {
 				return;
 			}
 
+			const planPath = cartItemForPlan?.product_slug
+				? getPlanPath( cartItemForPlan.product_slug )
+				: '';
+			const checkoutUrlWithArgs = addQueryArgs(
+				{ ...( withDiscount && { coupon: withDiscount } ) },
+				`/checkout/${ siteSlug }/${ planPath }`
+			);
+
 			page( checkoutUrlWithArgs );
 		},
-		[
-			isPlanUpsellEnabledOnFreeDomain.result,
-			onUpgradeClick,
-			paidDomainName,
-			siteSlug,
-			toggleIsFreePlanPaidDomainDialogOpen,
-			withDiscount,
-		]
+		[ onUpgradeClick, resolveDisplayedModal, siteSlug, withDiscount ]
 	);
 
 	const term = usePlanBillingPeriod( {
@@ -651,52 +639,28 @@ const PlansFeaturesMain = ( {
 			<QuerySites siteId={ siteId } />
 			<QuerySitePlans siteId={ siteId } />
 			<QueryActivePromotions />
-			{ paidDomainName && isFreePlanPaidDomainDialogOpen && (
-				<FreePlanPaidDomainDialog
+			{ lastClickedPlan && flowName && (
+				<ModalContainer
+					isModalOpen={ isModalOpen }
 					paidDomainName={ paidDomainName }
+					selectedPlan={ lastClickedPlan }
 					wpcomFreeDomainSuggestion={ wpcomFreeDomainSuggestion }
-					suggestedPlanSlug={ PLAN_PERSONAL }
-					isCustomDomainAllowedOnFreePlan={ isCustomDomainAllowedOnFreePlan }
-					onClose={ toggleIsFreePlanPaidDomainDialogOpen }
+					flowName={ flowName }
+					onClose={ () => setIsModalOpen( false ) }
 					onFreePlanSelected={ () => {
-						if ( isCustomDomainAllowedOnFreePlan.isLoading ) {
-							return;
-						}
-
-						if ( ! isCustomDomainAllowedOnFreePlan.result ) {
-							removePaidDomain?.();
-						}
+						removePaidDomain?.();
 						// Since this domain will not be available after it is selected, invalidate the cache.
 						invalidateDomainSuggestionCache();
 						wpcomFreeDomainSuggestion.result &&
 							setSiteUrlAsFreeDomainSuggestion?.( wpcomFreeDomainSuggestion.result );
 						onUpgradeClick?.( null );
 					} }
-					onPlanSelected={ () => {
-						const cartItemForPlan = getCartItemForPlan( PLAN_PERSONAL );
-						const cartItems = cartItemForPlan ? [ cartItemForPlan ] : null;
-						onUpgradeClick?.( cartItems );
-					} }
-				/>
-			) }
-			{ isFreePlanFreeDomainDialogOpen && (
-				<FreePlanFreeDomainDialog
-					suggestedPlanSlug={ PLAN_PERSONAL }
-					freeSubdomain={ resolvedSubdomainName }
-					onClose={ () => setIsFreePlanFreeDomainDialogOpen( false ) }
-					onFreePlanSelected={ () => {
+					onPlanSelected={ ( planSlug ) => {
 						if ( ! signupFlowSubdomain && wpcomFreeDomainSuggestion.result ) {
 							setSiteUrlAsFreeDomainSuggestion?.( wpcomFreeDomainSuggestion.result );
 						}
 						invalidateDomainSuggestionCache();
-						onUpgradeClick?.( null );
-					} }
-					onPlanSelected={ () => {
-						if ( ! signupFlowSubdomain && wpcomFreeDomainSuggestion.result ) {
-							setSiteUrlAsFreeDomainSuggestion?.( wpcomFreeDomainSuggestion.result );
-						}
-						invalidateDomainSuggestionCache();
-						const cartItemForPlan = getCartItemForPlan( PLAN_PERSONAL );
+						const cartItemForPlan = getCartItemForPlan( planSlug );
 						const cartItems = cartItemForPlan ? [ cartItemForPlan ] : null;
 						onUpgradeClick?.( cartItems );
 					} }
