+++ client/my-sites/plans-grid/components/features.tsx
@@ -29,8 +29,7 @@ const FreePlanCustomDomainFeature: React.FC< {
 	isCustomDomainAllowedOnFreePlan: DataResponse< boolean >;
 } > = ( { paidDomainName, wpcomFreeDomainSuggestion, isCustomDomainAllowedOnFreePlan } ) => {
 	const translate = useTranslate();
-	const isLoading =
-		wpcomFreeDomainSuggestion.isLoading || isCustomDomainAllowedOnFreePlan.isLoading;
+	const isLoading = wpcomFreeDomainSuggestion.isLoading;
 
 	return (
 		<SubdomainSuggestion>
