+++ packages/components/src/dialog/index.tsx
@@ -23,6 +23,8 @@ type Props = {
 	shouldCloseOnEsc?: boolean;
 	showCloseIcon?: boolean;
 	shouldCloseOnOverlayClick?: boolean;
+	labelledby?: string;
+	describedby?: string;
 };
 
 const Dialog = ( {
@@ -41,6 +43,8 @@ const Dialog = ( {
 	shouldCloseOnEsc,
 	showCloseIcon = false,
 	shouldCloseOnOverlayClick = true,
+	labelledby,
+	describedby,
 }: PropsWithChildren< Props > ) => {
 	const close = useCallback( () => onClose?.(), [ onClose ] );
 	const onButtonClick = useCallback(
@@ -66,6 +70,7 @@ const Dialog = ( {
 
 	return (
 		<Modal
+			aria={ { labelledby, describedby } }
 			isOpen={ isVisible }
 			onRequestClose={ close }
 			closeTimeoutMS={ leaveTimeout }
@@ -78,7 +83,11 @@ const Dialog = ( {
 			shouldCloseOnOverlayClick={ shouldCloseOnOverlayClick }
 		>
 			{ showCloseIcon && (
-				<button className="dialog__action-buttons-close" onClick={ () => onClose?.( this ) }>
+				<button
+					aria-label="Close"
+					className="dialog__action-buttons-close"
+					onClick={ () => onClose?.( this ) }
+				>
 					<Gridicon icon="cross" size={ 24 } />
 				</button>
 			) }
